requirements.txt
boto3
dotenv
pandas
psycopg2-binary
confluent-kafka
pytest